+++ vacasa/connect/connect.py
@@ -477,6 +477,70 @@ class VacasaConnect:
 
         return self._iterate_pages(url, headers, params)
 
+    def add_unit_amenity(self,
+                         amenity_id: int,
+                         unit_id: int,
+                         amenity_value: int = None,
+                         internal_notes: str = None,
+                         instructions: str = None,
+                         notes: str = None,
+                         provider_id: int = None) -> dict:
+        """Add a new unit amenity in Connect.
+
+        Args:
+            amenity_id: An amenity ID.
+            unit_id: A unit ID.
+            amenity_value: An amenity value.
+            internal_notes: An internal note.
+            instructions: An instruction.
+            notes: A note.
+            provider_id: A provider ID.
+
+        Returns:
+            json response for success
+        """
+        url = f"{self.endpoint}/v1/unit-amenities"
+        headers = self._headers()
+        payload = {
+            'data': {
+                'attributes': {
+                    'amenity_id': amenity_id,
+                    'unit_id': unit_id,
+                    'amenity_value': amenity_value,
+                    'internal_notes': internal_notes,
+                    'instructions': instructions,
+                    'notes': notes,
+                    'provider_id': provider_id,
+                }
+            }
+        }
+
+        r = self._post(url, json=payload, headers=headers)
+        r.raise_for_status()
+        return r.json()
+
+    def update_unit_amenity(self, unit_amenity_id, params: dict):
+        """Update a unit amenity in Connect.
+
+        Args:
+             unit_amenity_id: ID of the Unit Amenity to update.
+             params: A dict of key value pairs to update.
+
+        Yields:
+            An updated unit amenity.
+        """
+        url = f"{self.endpoint}/v1/unit-amenities/{unit_amenity_id}"
+        headers = self._headers()
+        payload = {
+            'data': {
+                'attributes': params
+            }
+        }
+
+        r = self._patch(url, json=payload, headers=headers)
+        r.raise_for_status()
+        return r.json()
+
     def get_unit_amenity_properties(self, params: dict = None):
         """Retrieve a list of all unit amenity properties
 
@@ -507,14 +571,20 @@ class VacasaConnect:
         url = f"{self.endpoint}/v1/unit-amenity-properties"
         headers = self._headers()
         payload = {
-            'amenity_id': amenity_id,
-            'unit_id': unit_id,
-            'idamenities_property': idamenities_property,
-            'property_value': property_value
+            'data': {
+                'type': 'unit-amenity-property',
+                'attributes': {
+                    'amenity_id': amenity_id,
+                    'unit_id': unit_id,
+                    'idamenities_property': idamenities_property,
+                    'property_value': property_value
+                }
+            }
         }
 
-        return self._post(url, json={'data': {'attributes': payload,
-                                              'type': 'unit-amenity-property'}}, headers=headers).json()
+        r = self._post(url, json=payload, headers=headers)
+        r.raise_for_status()
+        return r.json()
 
     def update_unit_amenity_property(self, unit_amenity_property_id, params: dict):
         """Update a unit amenity property in Connect.
@@ -528,7 +598,17 @@ class VacasaConnect:
         """
 
         url = f"{self.endpoint}/v1/unit-amenity-properties/{unit_amenity_property_id}"
-        return self._patch(url, json={'data': {'attributes': params}}, headers=self._headers()).json()
+        headers = self._headers()
+        payload = {
+            'data': {
+                'type': 'unit-amenity-property',
+                'attributes': params
+            }
+        }
+
+        r = self._patch(url, json=payload, headers=headers)
+        r.raise_for_status()
+        return r.json()
 
     def get_cities(self, params: dict = None):
         """Retrieve a list of all cities